added jwt authentication check

+++ resources/item.py
@@ -1,5 +1,6 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -11,14 +12,19 @@ blp = Blueprint("items", __name__, description="Operations on items")
 
 @blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
-        item = ItemModel.query.get_or_404(item_id)
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
 
+        item = ItemModel.query.get_or_404(item_id)
         try:
             db.session.delete(item)
             db.session.commit()
@@ -27,6 +33,7 @@ class Item(MethodView):
 
         return {"message": "Item deleted."}
 
+    @jwt_required()
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -45,6 +52,7 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         items = ItemModel.query.all()
@@ -53,6 +61,7 @@ class ItemList(MethodView):
     # @blp.arguments(ItemSchema) validates the JOSN data from client
     # and it calls this post method
     # and passes the validated data through item_data parameter
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
